keep sidebars open when you change pages
