--- src/components/NewSidebar/styles.scss
@@ -1,89 +1,0 @@
-@import '../../styles/colors';
-
-.toc-section {
-  ol {
-    list-style-type: none;
-    li {
-      line-height: 24px;
-      position: relative;
-      a {
-        color: $blue-black;
-        display: block;
-        font-weight: bold;
-      }
-    }
-  }
-  .menu-level-1 > li,
-  .menu-level-2 > li {
-    border-bottom: 1px solid $grey-hr;
-    border-left: 2px solid transparent;
-    > a {
-      font-size: 14px;
-      padding: 6px 16px;
-      &:hover {
-        background-color: $white;
-      }
-    }
-    &.link-active {
-      border-left-color: $magenta-border;
-      background-color: $white;
-    }
-  }
-  .menu-level-1 > li.menu-heading {
-    border-bottom-width: 0;
-    &:not(:first-child) {
-      margin-top: 20px;
-    }
-    h4 {
-      font-size: 13px;
-      font-weight: normal;
-      letter-spacing: 0.25px;
-      padding: 6px 16px;
-      text-transform: uppercase;
-    }
-  }
-  .menu-level-2 > li {
-    &.menu-active {
-      border-left-color: $magenta-border;
-      background-color: $white;
-      a:hover {
-        background-color: $grey-bg-hover;
-      }
-    }
-    &.menu-open {
-      border-left-color: $grey-border;
-    }
-  }
-  button.menu-toggle-btn {
-    background-color: transparent;
-    border: 0 none;
-    // border: 1px solid cyan;
-    cursor: pointer;
-    height: 36px;
-    padding-top: 4px;
-    position: absolute;
-    right: 0;
-    top: 0;
-    width: 36px;
-  }
-  .menu-level-3 > li {
-    > a {
-      font-size: 13px;
-      padding: 4px 24px;
-    }
-    button.menu-toggle-btn {
-      width: 32px;
-      height: 32px;
-    }
-  }
-  .menu-level-4 > li {
-    > a {
-      font-size: 12px;
-      font-weight: normal;
-      padding: 3px 8px 3px 40px;
-    }
-  }
-  .link-active > a {
-    color: $magenta;
-  }
-}
