+++ src/components/index.js
@@ -19,7 +19,6 @@ export { default as MarketingSection } from './MarketingSection';
 export { default as MegaMenu } from './NavBar/MegaMenu';
 export { default as NavBar } from './NavBar';
 export { default as NavLink } from './NavBar/NavLink';
-export { default as NewSidebar } from './NewSidebar';
 export { default as NoteBox } from './NoteBox';
 export { ProductFeature, ProductFeatureRow } from './ProductFeature';
 export { default as ProductHero } from './ProductHero';
